Repo: move the store to app folder
